move Mongo entity
